+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.24.0] - 2025-04-07
+
+### Added
+
+- Added support for overriding highlight options when highlighting in Polyline and PolylineCollection.
+- Added support for setting polyline options in the PolylineCollection.
+
+### Changed
+
+- Changed MarkerCollection and PolylineCollection handling of tags. Tags should now be a single string or an array of tags. Also updated the methods that pass “map” to have the tags first and map second.
+  
 ## [0.23.0] - 2025-03-24
 
 ## Fixed